+++ lend/src/components/ActionMMModal/Footer.tsx
@@ -1,0 +1,328 @@
+import React from "react";
+import { Col, Row } from "react-bootstrap";
+import colors from "shared/lib/designSystem/colors";
+import styled, { keyframes } from "styled-components";
+import { components } from "../../designSystem/components";
+import { PrimaryText, Title } from "../../designSystem";
+import sizes from "../../designSystem/sizes";
+import { getEtherscanURI, VaultOptions } from "../../constants/constants";
+import { formatBigNumber } from "../../utils/math";
+import { getUtilizationDecimals } from "../../utils/asset";
+import { usePoolsAPR } from "../../hooks/usePoolsAPR";
+import { fadeIn } from "shared/lib/designSystem/keyframes";
+import { css } from "styled-components";
+import { BaseUnderlineLink } from "shared/lib/designSystem";
+import { useVaultsData } from "../../hooks/web3DataContext";
+import { useWeb3React } from "@web3-react/core";
+import ExternalLinkIcon from "../Common/ExternalLinkIcon";
+import UtilizationBar from "../Common/UtilizationBar";
+import currency from "currency.js";
+import { ActionModalEnum } from ".";
+
+export const fadeInDisabled = keyframes`
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 0.64;
+  }
+`;
+
+export const FixedContainer = styled.div`
+  position: fixed;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: black;
+  z-index: 100;
+  width: 100%;
+  height: 100%;
+`;
+
+export const HeaderRow = styled(Row)`
+  height: ${components.header}px;
+  border-bottom: 1px solid ${colors.border};
+  z-index: 1;
+  margin-left: 0px;
+
+  > * {
+    padding: 0;
+
+    &:not(:last-child) {
+      border-right: 1px solid ${colors.border};
+    }
+  }
+`;
+
+export const FooterRow = styled(Row)`
+  min-height: ${components.footer}px;
+  border-top: 1px solid ${colors.border};
+  box-sizing: content-box;
+`;
+
+export const HeaderContainer = styled.div`
+  display: flex;
+  margin: auto;
+  width: 100%;
+  justify-content: center;
+  border-radius: 0;
+  min-height: ${components.header}px;
+  border-bottom: 1px solid ${colors.border};
+  z-index: 1;
+  color: ${colors.primaryText};
+
+  > * {
+    margin: auto 0;
+
+    &:not(:last-child) {
+      margin-right: 8px;
+    }
+  }
+`;
+
+export const HeaderText = styled(Title)`
+  font-size: 16px;
+  line-height: 20px;
+`;
+
+export const DetailContainer = styled.div<{ show?: boolean; delay?: number }>`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  height: 100%;
+  ${({ show, delay }) => {
+    return (
+      show &&
+      css`
+        opacity: 0;
+
+        &:disabled {
+          opacity: 0;
+        }
+
+        animation: ${fadeIn} 1s ease-in-out forwards;
+        animation-delay: ${delay || 0}s;
+      `
+    );
+  }}
+`;
+
+export const Content = styled(Row)`
+  height: calc(100% - ${components.header}px - ${components.footer}px);
+
+  @media (max-width: ${sizes.lg}px) {
+    height: 100%;
+  }
+
+  > *:not(:last-child) {
+    border-right: 1px solid ${colors.border};
+  }
+`;
+
+const DesktopOnly = styled.div`
+  display: flex;
+  width: 100%;
+  > * {
+    padding: 0;
+
+    &:not(:last-child) {
+      border-right: 1px solid ${colors.border};
+    }
+  }
+  @media (max-width: ${sizes.lg}px) {
+    display: none;
+  }
+`;
+
+const MobileOnly = styled.div`
+  display: none;
+  width: 100%;
+  > * {
+    padding: 0;
+
+    &:not(:last-child) {
+      border-right: 1px solid ${colors.border};
+    }
+  }
+  @media (max-width: ${sizes.lg}px) {
+    display: flex;
+  }
+`;
+
+const DetailTitle = styled.div`
+  font-size: 12px;
+  color: ${colors.primaryText}52;
+  text-align: center;
+`;
+
+export const DetailText = styled(Title)<{ color?: string }>`
+  font-size: 14px;
+  line-height: 20px;
+  color: ${(props) => (props.color ? props.color : ``)};
+`;
+
+export const StyledPrimaryText = styled(PrimaryText)`
+  font-size: 14px;
+  line-height: 20px;
+  text-decoration: underline;
+  margin-right: 4px;
+`;
+
+export const TransactionContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  height: 100%;
+  width: 100%;
+`;
+
+const UnderlineLink = styled(BaseUnderlineLink)`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  width: 100%;
+`;
+interface FooterProps {
+  pool: VaultOptions;
+  page: ActionModalEnum;
+  show: boolean;
+  txhash: string | undefined;
+  borrowAmount: string;
+  isBorrow: boolean;
+}
+
+const Footer: React.FC<FooterProps> = ({
+  show,
+  pool,
+  page,
+  txhash,
+  borrowAmount,
+  isBorrow,
+}) => {
+  const vaultDatas = useVaultsData();
+  const { aprs } = usePoolsAPR();
+  const { chainId } = useWeb3React();
+  const apr = aprs[pool];
+  const utilizationDecimals = getUtilizationDecimals();
+  const utilizationRate = vaultDatas.data[pool].utilizationRate;
+  const poolSize = vaultDatas.data[pool].poolSize;
+  const borrowAmountNum = parseInt(borrowAmount);
+  const absoluteBorrowAmount = Math.abs(parseInt(borrowAmount));
+
+  return (
+    <>
+      {page === ActionModalEnum.PREVIEW ? (
+        <FooterRow>
+          <DesktopOnly>
+            <Col xs={3}>
+              <DetailContainer show={show} delay={0.1}>
+                <DetailTitle>Current Utilization</DetailTitle>
+                <div className="d-flex">
+                  <UtilizationBar
+                    percent={parseFloat(
+                      formatBigNumber(utilizationRate, utilizationDecimals)
+                    )}
+                    width={40}
+                    color="white"
+                  />
+                  <DetailText>
+                    {formatBigNumber(utilizationRate, utilizationDecimals)}%
+                  </DetailText>
+                </div>
+              </DetailContainer>
+            </Col>
+            <Col xs={3}>
+              <DetailContainer show={show} delay={0.2}>
+                <DetailTitle>Pool Size (USDC)</DetailTitle>
+                <DetailText>
+                  {currency(formatBigNumber(poolSize, 6), {
+                    symbol: "",
+                  }).format()}
+                </DetailText>
+              </DetailContainer>
+            </Col>
+            <Col xs={3}>
+              <DetailContainer show={show} delay={0.3}>
+                <DetailTitle>
+                  {isBorrow ? "Borrow" : "Repay"} Amount
+                </DetailTitle>
+                <DetailText
+                  color={
+                    borrowAmountNum === 0
+                      ? colors.primaryText
+                      : borrowAmountNum > 0
+                      ? colors.green
+                      : colors.red
+                  }
+                >
+                  {currency(
+                    absoluteBorrowAmount
+                      ? formatBigNumber(absoluteBorrowAmount, 6)
+                      : "0.00",
+                    {
+                      symbol: "",
+                    }
+                  ).format()}
+                </DetailText>
+              </DetailContainer>
+            </Col>
+            <Col xs={3}>
+              <DetailContainer show={show} delay={0.4}>
+                <DetailTitle>Borrow APR</DetailTitle>
+                <DetailText>
+                  {currency(apr.toFixed(2), { symbol: "" }).format()}%
+                </DetailText>
+              </DetailContainer>
+            </Col>
+          </DesktopOnly>
+          <MobileOnly>
+            <Col xs={6}>
+              <DetailContainer show={show} delay={0.1}>
+                <DetailTitle>Current Utilization</DetailTitle>
+                <div className="d-flex">
+                  <UtilizationBar
+                    percent={parseFloat(
+                      formatBigNumber(utilizationRate, utilizationDecimals)
+                    )}
+                    width={40}
+                    color="white"
+                  />
+                  <DetailText>
+                    {formatBigNumber(utilizationRate, utilizationDecimals)}%
+                  </DetailText>
+                </div>
+              </DetailContainer>
+            </Col>
+            <Col xs={6}>
+              <DetailContainer show={show} delay={0.2}>
+                <DetailTitle>Borrow APR</DetailTitle>
+                <DetailText>
+                  {currency(apr.toFixed(2), { symbol: "" }).format()}%
+                </DetailText>
+              </DetailContainer>
+            </Col>
+          </MobileOnly>
+        </FooterRow>
+      ) : (
+        <FooterRow>
+          {chainId !== undefined && (
+            <UnderlineLink
+              to={`${getEtherscanURI(chainId)}/tx/${txhash}`}
+              target="_blank"
+              rel="noreferrer noopener"
+            >
+              <StyledPrimaryText>View on Etherscan</StyledPrimaryText>
+              <ExternalLinkIcon />
+            </UnderlineLink>
+          )}
+        </FooterRow>
+      )}
+    </>
+  );
+};
+
+export default Footer;
